+++ src/main/java/builderb0y/autocodec/reflection/memberViews/FieldView.java
@@ -2,11 +2,9 @@ package builderb0y.autocodec.reflection.memberViews;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
-import java.lang.reflect.AnnotatedType;
 import java.lang.reflect.Field;
 import java.util.function.UnaryOperator;
 
-import org.jetbrains.annotations.ApiStatus.Internal;
 import org.jetbrains.annotations.NotNull;
 
 import builderb0y.autocodec.annotations.UseGetter;
@@ -153,8 +151,16 @@ public class FieldView<T_Owner, T_Member> extends FieldLikeMemberView<T_Owner, T
 	}
 
 	@Override
-	@Internal
-	public @NotNull AnnotatedType getAnnotatedType() {
-		return this.field.getAnnotatedType();
+	public @NotNull ReifiedType<T_Member> getType() {
+		ReifiedType<T_Member> type = this.type;
+		if (type == null) {
+			this.type = type = (
+				this
+				.getDeclaringType()
+				.resolveDeclaration(this.field.getAnnotatedType())
+				.uncheckedCast()
+			);
+		}
+		return type;
 	}
 }